% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeNormalize.R
\name{degreeNormalize}
\alias{degreeNormalize}
\title{FUNCTION Weight edges in adjaceny matrix by the degree of interacting vertices.}
\usage{
degreeNormalize(adj.mat)
}
\arguments{
\item{adj.mat}{Adjacency matrix.  The matrix needs should be from an undirected network 
and should be symmetric across the diagonal.}
}
\value{
degree normalized matrix
}
\description{
Takes as input an adjacency matrix and return as degree normalized matrix 
of the same dimensions.  The adjacency matrix should be from and undirected network
and thus should be symmetric.
}
\examples{
myMat = matrix(c(0, 1, 1, 1, 1, 0, 0, 1, 1, 0 , 0, 0, 1, 1, 0, 0), 4, 4)
myMat.norm = degreeNormalize(myMat)
myMat.norm

}
