% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.smoothing.R
\docType{package}
\name{network.smoothing}
\alias{network.smoothing}
\alias{network.smoothing-package}
\title{A package to propagate data over an interaction network}
\description{
This uses the method similar to network-based stratification (NBS) method
published in Nature Methods (2013)
}
\details{
Pilot version of function
}
